% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dairy_region_map}
\alias{dairy_region_map}
\title{Simple feature collection with 5 fields}
\format{
\subsection{\code{dairy_region_map}}{

\describe{
\item{region_name}{Region name}
\item{geometry}{geometry}
\item{is_dairy}{Logical denoting whether region is a dairy region}
\item{cent_long}{Longitude of centroid}
\item{cent_lat}{Latitude of centroid}
...
}
}
}
\source{
\url{https://www.vic.gov.au/regional-model-department-education}
}
\usage{
dairy_region_map
}
\description{
Simple feature collection with 5 fields
}
\keyword{datasets}
