% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dairy_region_tbl}
\alias{dairy_region_tbl}
\title{Region Table}
\format{
\subsection{\code{region_tbl}}{

\describe{
\item{region_name}{Region name, with four values}
\item{lga_name}{Local government area name, with 79 values}
\item{is_dairy}{Logical denoting whether region is a dairy region}
...
}
}
}
\source{
\url{https://www.vic.gov.au/regional-model-department-education}
}
\usage{
dairy_region_tbl
}
\description{
A tibble of with 3 columns and 79 rows.
}
\keyword{datasets}
